use tanh for float

+++ source/lib/src/cuda/gelu.cu
@@ -2,7 +2,7 @@
 #include "device.h"
 
 __device__ inline double _tanh(double x) {return tanh(x);}
-__device__ inline float _tanh(float x) {return tanh(x);}
+__device__ inline float _tanh(float x) {return tanhf(x);}
 
 template <typename FPTYPE>
 __global__ void gelu(
